+++ src/main/java/org/folio/inventoryupdate/InventoryQuery.java
@@ -6,10 +6,6 @@ import java.nio.charset.StandardCharsets;
 public abstract class InventoryQuery {
     protected String queryString;
 
-    public String getQueryString( ) {
-        return queryString;
-    }
-
     public String getURLEncodedQueryString( ) {
         return URLEncoder.encode(queryString, StandardCharsets.UTF_8);
     }